Moved the XXXd stubs into dedicated files
